# frontend/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI焦化厂智能化平台 API",
    "description": "智能配煤 / 焦炉加热控制 / 设备PdM / 安全视觉 四大应用后端",
    "version": "1.1.0"
  },
  "paths": {
...
    "/blend/optimize": {
      "post": {
        "tags": [
          "智能配煤"
        ],
        "summary": "调配煤优化器",
        "description": "运行配煤优化：GA(scikit-opt)寻优 + 质量预测模型评估 + 边际贡献解释。\n\n返回3套方案：cost_optimal / quality_stable / balanced（4.1.5输出规范）。\n优化器本体在 models/blending_optimizer/，此处只做参数校验与调用编排。",
        "operationId": "optimize_blend_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/blend/recipes": {
      "get": {
        "tags": [
          "智能配煤"
        ],
        "summary": "历史配煤方案查询",
        "description": "查询PostgreSQL blend_recipe表（3.4.1节），支持按焦炉号过滤。",
        "operationId": "list_recipes_blend_recipes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "furnace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Furnace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/blend/feedback": {
      "post": {
        "tags": [
          "智能配煤"
        ],
        "summary": "化验结果回流，触发增量学习",
        "description": "化验数据闭环（4.1.7节）：\n\n1. 取该批次当时的预测值 → 2. 计算误差 → 3. 超阈值写入误差缓冲区(Redis)\n→ 4. 缓冲区满 BUFFER_TRIGGER_SIZE 条触发增量微调（LightGBM热启动，小学习率防遗忘）。",
        "operationId": "feedback_blend_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/furnace/temp": {
      "get": {
        "tags": [
          "焦炉加热控制"
        ],
        "summary": "实时炉温",
        "description": "查询TDengine中该焦炉最近一次的机/焦侧火道温度。\n\n注意：换向前后2~3分钟数据为脏数据（4.2.6节），\n由ETL管道打标记，此处查询应过滤 switching_flag=0。",
        "operationId": "get_temp_furnace_temp_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "furnace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Furnace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/furnace/ai-setpoint": {
      "get": {
        "tags": [
          "焦炉加热控制"
        ],
        "summary": "AI设定值建议",
        "description": "MPC求解最优煤气流量/吸力设定值（4.2.4节），输出前经 clamp_setpoint 钳位。\n\nshadow模式下仅展示不下发；auto模式由控制回路消费本接口结果。",
        "operationId": "get_ai_setpoint_furnace_ai_setpoint_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "furnace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Furnace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/furnace/control-mode": {
      "post": {
        "tags": [
          "焦炉加热控制"
        ],
        "summary": "切换控制模式",
        "description": "切换 manual/shadow/auto。\n\n状态存Redis（key: furnace:control_mode:{furnace_id}），Redis不可用时落内存。\n注意：「一键切手动」为DCS硬切换（4.2.6节），本接口只是AI系统侧的模式记录，\n不构成安全通路。",
        "operationId": "set_control_mode_furnace_control_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/furnace/k-coefficients": {
      "get": {
        "tags": [
          "焦炉加热控制"
        ],
        "summary": "热工K系数查询",
        "description": "返回K均/K安/K1/K2/K3（4.2.6节统一考核口径）。\n\n计算逻辑在 data/pipeline/k_coefficients.py（直行温度记录+推焦计划/执行记录），\n本接口只做查询编排。",
        "operationId": "get_k_coefficients_furnace_k_coefficients_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "furnace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Furnace Id"
            }
          },
          {
            "name": "shift_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shift Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlendSolution": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "方案类型：cost_optimal/quality_stable/balanced"
          },
          "blend_ratio": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Blend Ratio",
            "description": "各煤种配比（和为1）"
          },
          "estimated_cost": {
            "type": "number",
            "title": "Estimated Cost",
            "description": "预估吨煤成本 元/吨"
          },
          "predicted_quality": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Predicted Quality",
            "description": "预测质量：M25/M10/CSR/CRI"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "description": "置信度 0~1"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "description": "SHAP/边际贡献中文解释",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "type",
          "blend_ratio",
          "estimated_cost",
          "predicted_quality",
          "confidence"
        ],
        "title": "BlendSolution",
        "description": "单套配煤方案（方案§4.1.5 输出规范）。\n\n字段与 models/blending_optimizer/optimizer.py 的 BlendSolution 对齐，\n此处为服务层契约（不依赖 scikit-opt 重依赖，可被路由轻量引用）。"
      },
      "CoalInfo": {
        "properties": {
          "coal_id": {
            "type": "integer",
            "title": "Coal Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "stock_tons": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Stock Tons",
            "description": "库存量 吨"
          },
          "price_per_ton": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price Per Ton",
            "description": "单价 元/吨"
          },
          "min_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Ratio",
            "description": "配比下限（工艺约束）",
            "default": 0.0
          },
          "max_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Ratio",
            "description": "配比上限（工艺约束）",
            "default": 1.0
          },
          "lab_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Lab Data",
            "description": "化验指标：Ad/Vdaf/St_d/G/Y/Rmax等"
          }
        },
        "type": "object",
        "required": [
          "coal_id",
          "name",
          "stock_tons",
          "price_per_ton"
        ],
        "title": "CoalInfo",
        "description": "可用煤种信息（含库存约束与化验指标）。"
      },
      "ControlModeRequest": {
        "properties": {
          "furnace_id": {
            "type": "integer",
            "title": "Furnace Id",
            "description": "焦炉编号"
          },
          "mode": {
            "type": "string",
            "pattern": "^(shadow|auto)$",
            "title": "Mode",
            "description": "仅允许影子/自动软切换"
          },
          "operator": {
            "type": "string",
            "minLength": 1,
            "title": "Operator",
            "description": "操作人工号/姓名"
          },
          "reviewer": {
            "type": "string",
            "minLength": 1,
            "title": "Reviewer",
            "description": "复核人工号/姓名"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason",
            "description": "切换原因"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "description": "复核人现场确认后必须为 true"
          }
        },
        "type": "object",
        "required": [
          "furnace_id",
          "mode",
          "operator",
          "reviewer",
          "reason",
          "confirm"
        ],
        "title": "ControlModeRequest",
        "description": "AI控制模式软切换请求（方案§4.2.2/§4.2.6）。"
      },
      "ControlModeResponse": {
        "properties": {
          "furnace_id": {
            "type": "integer",
            "title": "Furnace Id",
            "description": "焦炉编号"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "切换后的控制模式"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "状态：ok"
          }
        },
        "type": "object",
        "required": [
          "furnace_id",
          "mode",
          "status"
        ],
        "title": "ControlModeResponse",
        "description": "POST /furnace/control-mode 响应（方案§4.2.2）。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LabFeedback": {
        "properties": {
          "batch_no": {
            "type": "string",
            "title": "Batch No",
            "description": "配煤批次号"
          },
          "actual_quality": {
            "additionalProperties": true,
            "type": "object",
            "title": "Actual Quality",
            "description": "实测焦炭质量：M25/M10/CSR/CRI"
          }
        },
        "type": "object",
        "required": [
          "batch_no",
          "actual_quality"
        ],
        "title": "LabFeedback",
        "description": "化验结果回流（4.1.7增量学习入口）。"
      },
...
      "OptimizeRequest": {
        "properties": {
          "target_quality": {
            "$ref": "#/components/schemas/TargetQuality"
          },
          "available_coals": {
            "items": {
              "$ref": "#/components/schemas/CoalInfo"
            },
            "type": "array",
            "minItems": 2,
            "title": "Available Coals"
          },
          "max_cost_per_ton": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Cost Per Ton",
            "default": 1500.0
          },
          "priority": {
            "type": "string",
            "pattern": "^(cost|quality|balanced)$",
            "title": "Priority",
            "default": "cost"
          }
        },
        "type": "object",
        "required": [
          "target_quality",
          "available_coals"
        ],
        "title": "OptimizeRequest",
        "description": "配煤优化请求（4.1.5输入JSON）。"
      },
      "OptimizeResponse": {
        "properties": {
          "solutions": {
            "items": {
              "$ref": "#/components/schemas/BlendSolution"
            },
            "type": "array",
            "title": "Solutions",
            "description": "三套方案"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "模型版本"
          },
          "compute_time_ms": {
            "type": "integer",
            "title": "Compute Time Ms",
            "description": "求解耗时 毫秒",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "solutions",
          "model_version"
        ],
        "title": "OptimizeResponse",
        "description": "配煤优化响应（方案§4.1.5）。"
      },
...
      "RecipeListResponse": {
        "properties": {
          "recipes": {
            "items": {
              "$ref": "#/components/schemas/RecipeRow"
            },
            "type": "array",
            "title": "Recipes"
          }
        },
        "type": "object",
        "required": [
          "recipes"
        ],
        "title": "RecipeListResponse",
        "description": "GET /blend/recipes 响应。"
      },
      "RecipeRow": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "batch_no": {
            "type": "string",
            "title": "Batch No"
          },
          "furnace_id": {
            "type": "integer",
            "title": "Furnace Id"
          },
          "production_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Production Date"
          },
          "total_coal_tons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Coal Tons"
          },
          "estimated_cost_per_ton": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Per Ton"
          },
          "actual_cost_per_ton": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Cost Per Ton"
          },
          "ai_generated": {
            "type": "boolean",
            "title": "Ai Generated",
            "default": false
          },
          "operator_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_no",
          "furnace_id"
        ],
        "title": "RecipeRow",
        "description": "历史配煤方案行（blend_recipe 表，方案§3.4.1）。"
      },
      "TargetQuality": {
        "properties": {
          "M25_min": {
            "type": "number",
            "title": "M25 Min",
            "description": "抗碎强度M25下限 %",
            "default": 90.0
          },
          "M10_max": {
            "type": "number",
            "title": "M10 Max",
            "description": "耐磨强度M10上限 %",
            "default": 6.5
          },
          "CSR_min": {
            "type": "number",
            "title": "Csr Min",
            "description": "反应后强度CSR下限 %",
            "default": 65.0
          },
          "CRI_max": {
            "type": "number",
            "title": "Cri Max",
            "description": "反应性CRI上限 %",
            "default": 25.0
          }
        },
        "type": "object",
        "title": "TargetQuality",
        "description": "目标焦炭质量指标。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}